requirements.txt
streamlit>=1.43
pandas>=2.2
numpy>=1.26
plotly>=5.24
scikit-learn>=1.4
xgboost>=2.1
lightgbm>=4.5
python-dotenv>=1.0
shap>=0.46
requests>=2.32
